+++ SocketNetworking/Server/NetworkServer.cs
@@ -227,7 +227,7 @@ namespace SocketNetworking.Server
                 NetworkClient cursedClient = (NetworkClient)Convert.ChangeType(client, ClientType);
                 _clients.Add(clientId, cursedClient);
                 ClientHandler handler = handlers.Next();
-                handler.AddClient(cursedClient);
+                //handler.AddClient(cursedClient);
                 Log.GlobalDebug($"Added client. ID: {clientId}, Type: {cursedClient.GetType().FullName}");
             }
         }
@@ -423,87 +423,7 @@ namespace SocketNetworking.Server
     }
 
 
-    public class ClientHandler : IRoundRobinData<ClientHandler>
-    {
-        public Thread Thread { get; }
 
-        /// <summary>
-        /// Not thread safe.
-        /// </summary>
-        public HashSet<NetworkClient> Clients { get; }
-
-        public ClientHandler(IEnumerable<NetworkClient> clients)
-        {
-            Thread = new Thread(Run);
-            Clients = new HashSet<NetworkClient>(clients);
-        }
-
-        public int CurrentClientCount
-        {
-            get
-            {
-                return Clients.Count;
-            }
-        }
-
-        public bool AllowChoosing => CurrentClientCount < NetworkServer.Config.ClientsPerThread;
-
-        public bool AllowSorting => true;
-
-        public void AddClient(NetworkClient client)
-        {
-            lock(_lock)
-            {
-                Clients.Add(client);
-            }
-        }
-
-        public void RemoveClient(NetworkClient client)
-        {
-            lock(_lock)
-            {
-                Clients.Remove(client);
-            }
-        }
-
-        public ClientHandler() : this(new List<NetworkClient>()) { }
-
-        bool die = false;
-
-        public void Start()
-        {
-            Thread.Start();
-        }
-
-        public void Stop()
-        {
-            die = true;
-            Thread.Abort();
-        }
-
-        object _lock = new object();
-
-        void Run()
-        {
-            while(true)
-            {
-                if (die) return;
-                lock (_lock)
-                {
-                    foreach (NetworkClient client in Clients)
-                    {
-                        client.ReadNext();
-                        client.WriteNext();
-                    }
-                }
-            }
-        }
-
-        public int CompareTo(ClientHandler other)
-        {
-            return CurrentClientCount - other.CurrentClientCount;
-        }
-    }
 
     public enum ServerState 
     {
